+++ example/actions/index.ts
@@ -1,0 +1,5 @@
+import { onChat } from "./chat";
+import { onPing } from "./ping";
+import { onSubscribe } from "./subscribe";
+
+export const actions = [onChat, onPing, onSubscribe];