+++ src/attach.spec.ts
@@ -50,7 +50,12 @@ describe("Attach", () => {
     };
     const loggerMock = { info: vi.fn(), debug: vi.fn() };
     const actionsMock = [
-      { execute: vi.fn(), getNamespace: () => "/", getEvent: () => "test" },
+      {
+        execute: vi.fn(),
+        getNamespace: () => "/",
+        getEvent: () => "test",
+        getSchema: vi.fn(),
+      },
     ];
 
     test("should set the listeners", async () => {